* issues/858/events/application:
  chore: review feedback improvements
  chore: review feedback improvments
  refactor: testevents.ResetEventManager() to return cosmos sdk as well
  fix: bugs & shared helper return type

+++ x/tokenomics/keeper/settle_pending_claims.go
@@ -263,7 +263,7 @@ func (k Keeper) getExpiringClaims(ctx sdk.Context) (expiringClaims []prooftypes.
 	// expiringSessionEndHeight is the session end height of the session whose proof
 	// window has most recently closed.
 	sessionEndToProofWindowCloseNumBlocks := sharedtypes.GetSessionEndToProofWindowCloseBlocks(sharedParams)
-	expiringSessionEndHeight := blockHeight - int64(sessionEndToProofWindowCloseNumBlocks+1)
+	expiringSessionEndHeight := blockHeight - (sessionEndToProofWindowCloseNumBlocks + 1)
 
 	var nextKey []byte
 	for {
